build: Start a novel iteration of cleopatra
So, this is fun. We start implementing the next version of cleopatra,
with a slightly different approach: this time, we implement a tool in
Rust rather than relying solely on Make.

This next iteration of cleopatra is implemented in Rust, and for now,
the “current” version of cleopatra is used as the build system. It
will be an interesting experiment to bootstrap the process once again:
that is, start using the Rust tool to build itself.